pyproject.toml
[project]
name = "searchable-pexels"
version = "0.1.0"
description = "Search through video datasets by caption (fuzzy) and visual similarity (CLIP)"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "flask-cors>=4.0",
    "fire>=0.5",
    "numpy>=1.24",
    "Pillow>=10.0",
    "transformers>=4.36",
    "torch>=2.0",
    "faiss-cpu>=1.7",
    "tqdm>=4.65",
    "rp",
    "opencv-python>=4.8",
    "opencv-contrib-python>=4.8",
    "requests>=2.28",
    "torchvision>=0.15",
    "py3nvml>=0.2",
    "easydict>=1.9",
    "einops>=0.6",
    "imagehash>=4.3",
    "decord2>=3.0",
    "psutil>=7.2.2",
    "pyinstrument>=5.1.2",
    "av>=16.1.0",
    "pynvvideocodec>=2.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["server", "preprocess"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
